Rename module tx_pool to pool.
In anticipation of creating tx, batch and block pools.
